requirements.txt
Flask==1.1.1
PyJWT==1.7.1
requests==2.21.0